+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * [bugfix] Do not show `shelly deploys show last` instruction if last deployment was made by admin and was failed.
 * [improvement] There is no longer configuration_failed state for apps.
+* [bugfix] Do not create new branch after running `shelly setup`
 
 ## 0.4.2 / 2013-08-29